refactor signup and login files
