+++ src/interface_adapter/login/LoginState.java
@@ -1,24 +1,20 @@
-package interface_adapter;
+package interface_adapter.login;
 
-public class SignupState {
+public class LoginState {
     private String username = "";
     private String usernameError = null;
     private String password = "";
     private String passwordError = null;
-    private String repeatPassword = "";
-    private String repeatPasswordError = null;
 
-    public SignupState(SignupState copy) {
+    public LoginState(LoginState copy) {
         username = copy.username;
         usernameError = copy.usernameError;
         password = copy.password;
         passwordError = copy.passwordError;
-        repeatPassword = copy.repeatPassword;
-        repeatPasswordError = copy.repeatPasswordError;
     }
 
     // Because of the previous copy constructor, the default constructor must be explicit.
-    public SignupState() {}
+    public LoginState() {}
 
     public String getUsername() {
         return username;
@@ -36,14 +32,6 @@ public class SignupState {
         return passwordError;
     }
 
-    public String getRepeatPassword() {
-        return repeatPassword;
-    }
-
-    public String getRepeatPasswordError() {
-        return repeatPasswordError;
-    }
-
     public void setUsername(String username) {
         this.username = username;
     }
@@ -58,13 +46,5 @@ public class SignupState {
 
     public void setPasswordError(String passwordError) {
         this.passwordError = passwordError;
-    }
-
-    public void setRepeatPassword(String repeatPassword) {
-        this.repeatPassword = repeatPassword;
-    }
-
-    public void setRepeatPasswordError(String repeatPasswordError) {
-        this.repeatPasswordError = repeatPasswordError;
     }
 }
