+++ src/interface_adapter/signup/SignupPresenter.java
@@ -1,5 +1,8 @@
-package interface_adapter;
+package interface_adapter.signup;
 
+import interface_adapter.login.LoginState;
+import interface_adapter.login.LoginViewModel;
+import interface_adapter.ViewManagerModel;
 import use_case.signup.SignupOutputBoundary;
 import use_case.signup.SignupOutputData;
