+++ src/interface_adapter/signup/SignupState.java
@@ -1,20 +1,24 @@
-package interface_adapter;
+package interface_adapter.signup;
 
-public class LoginState {
+public class SignupState {
     private String username = "";
     private String usernameError = null;
     private String password = "";
     private String passwordError = null;
+    private String repeatPassword = "";
+    private String repeatPasswordError = null;
 
-    public LoginState(LoginState copy) {
+    public SignupState(SignupState copy) {
         username = copy.username;
         usernameError = copy.usernameError;
         password = copy.password;
         passwordError = copy.passwordError;
+        repeatPassword = copy.repeatPassword;
+        repeatPasswordError = copy.repeatPasswordError;
     }
 
     // Because of the previous copy constructor, the default constructor must be explicit.
-    public LoginState() {}
+    public SignupState() {}
 
     public String getUsername() {
         return username;
@@ -32,6 +36,14 @@ public class LoginState {
         return passwordError;
     }
 
+    public String getRepeatPassword() {
+        return repeatPassword;
+    }
+
+    public String getRepeatPasswordError() {
+        return repeatPasswordError;
+    }
+
     public void setUsername(String username) {
         this.username = username;
     }
@@ -46,5 +58,13 @@ public class LoginState {
 
     public void setPasswordError(String passwordError) {
         this.passwordError = passwordError;
+    }
+
+    public void setRepeatPassword(String repeatPassword) {
+        this.repeatPassword = repeatPassword;
+    }
+
+    public void setRepeatPasswordError(String repeatPasswordError) {
+        this.repeatPasswordError = repeatPasswordError;
     }
 }
